the problem to sharing state between jest test files solved via a data layer inside test layer that reads and writes to a .json file

+++ src/db/db-models.provider.ts
@@ -2,19 +2,26 @@ import { Mongoose } from 'mongoose';
 import { UserSchema } from './schemas/user.schema';
 
 let models: Record<string, Mongoose['Model']>;
+export function initModels(mongoose: Mongoose): void {
+    models = {
+        User: mongoose.model('User', UserSchema),
+    };
+}
+
 export const dbModelsProvider = {
     provide: 'DB_MODELS',
     useFactory: (mongoose: Mongoose): Record<string, Mongoose['Model']> => {
         console.log('creating models...');
-        models = {
-            User: mongoose.model('User', UserSchema),
-        };
+        initModels(mongoose);
         return models;
     },
     inject: ['DB_CONNECTION'],
 };
 
-export function cleanDb() {
-    models.User.deleteMany({});
+export async function cleanDb(): Promise<void> {
+    if (!models || !models.User) {
+        throw new Error('Models not initialized');
+    }
+    await models.User.deleteMany({});
     console.log(`\ndb cleaned\n`);
 }
